Adjust fixed-pitched faces for org-mode
These faces were too big with the default fixed-pitch font and that got out of
sync with the look and feel of the theme.

There was no need to adjust the height of `org-date`, it displayed fine in a
smaller resolution screen, and Org's source code blocks needed a few tweaks due
to them using the default font for the buffer, which is overriden by the
variable-pitch face in `variable-pitch-mode`.