Adjust evil mode-line faces
They were too bright in the light theme. The idea is to be able to
distinguish in which more you're in with the corner of your eyes (and by
the shape of the cursor), for that, I had to reduce the brightness of
the color indicators in the mode line.