Adjust colors for terminal-based modes
The colors inside shell command outputs and term-related modes is
controlled by the variable `ansi-color-map`. To set it, you need to
change the vector `ansi-color-names-vector` with the colors you want to
put there.

The order is:

black
red
yellow
green
blue
purple
cyan
gray

After that, we need to pass the product of the new color vector to
`ansi-color-map`  using `ansi-color-make-color-map`.